% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo_statistica_decisioni.R
\docType{data}
\name{camellia}
\alias{camellia}
\title{dataset CAMELLIA}
\format{\describe{
\item{id}{numero identificativo della foglia}
\item{lunghezza}{lunghezza del diametro maggiore della foglia in mm}
\item{larghezza}{larghezza del diametro minore della foglia in mm}
\item{pesovivo}{peso della foglia misurato 1 ora dopo la raccolta in deci-milligrammi}
\item{pesosecco}{peso della foglia misurato 40 giorni dopo la raccolta in deci-milligrammi}
\item{codice}{codice della specie, caratterizzante la forma: lanceolata=1; cordiforme=0}
}}
\source{
Piccolo, D., Statistica per le decisioni, Il Mulino 2010
}
\usage{
camellia
}
\description{
I dati riguardano n.333 foglie di Camellia Japonica. Le foglie sono
state raccolte, pesate, essiccate e ripesate dopo 40 giorni estivi
in appartamento
}
\examples{
head(camellia, n = 10)
summary(camellia)
}
\keyword{datasets}
