% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo_statistica_decisioni.R
\docType{data}
\name{compact}
\alias{compact}
\title{dataset COMPACT}
\format{\describe{
\item{durata}{durata del singolo brano in secondi}
\item{anno}{anno di edizione del brano}
}}
\source{
Piccolo, D., Statistica per le decisioni, Il Mulino 2010
}
\usage{
compact
}
\description{
I dati riguardano n.1192 brani incisi su compact disc di una
 collezione privata. Per ciascun brano e` nota sia la durata che
 l'anno di edizione
}
\examples{
head(compact, n = 10)
summary(compact)
}
\keyword{datasets}
