% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo_statistica.R
\docType{data}
\name{film}
\alias{film}
\title{dataset FILM}
\format{\describe{
\item{giudizio}{giudizio}
\item{genere}{genere}
\item{naziona}{nazione}
\item{anno}{anno}
\item{durata}{durata}
\item{visibilita}{visibilità}
}}
\source{
Piccolo, D., Statistica, Il Mulino 2000
}
\usage{
film
}
\description{
dataset film
}
\examples{
head(film, n = 10)
summary(film)
}
\keyword{datasets}
