% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frei.R
\docType{data}
\name{frei}
\alias{frei}
\title{Acute Leukemia data from Freireich et al (1963).}
\format{A data frame with 42 observations on the following variables.
 \describe{
   \item{\code{time}}{time in remission (in weeks)}
   \item{\code{status}}{event status, 1 is relapse, 0 is censored}
   \item{\code{treatment}}{treatment group: either 'placebo' or '6-MP'}
   \item{\code{pair}}{pair id number}
}}
\source{
bpcp package
}
\usage{
frei
}
\description{
In this study there were 21 pairs of subjects, and within each
 pair one subject received 6-mercaptopurine (6-MP) and one got
 placebo. The data are right censored.
}
\details{
See also Gehan (1965) who used the data ignoring the pairing so
 that he could illustrate his famous two-sample (non-paired) rank
 test.
}
\examples{
head(frei)
}
\references{
Gehan (1965) Biometrika 52:203-223.
Freireich et al (1963) Blood 21(6):699-716.
}
\keyword{datasets}

