% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\docType{data}
\name{libri}
\alias{libri}
\title{dataset PREZZOLIBRI}
\format{\describe{
\item{prezzo}{prezzo in lire del libro}
}}
\source{
Piccolo, D., Statistica per le decisioni, Il Mulino 2010
}
\usage{
libri
}
\description{
I dati riguardano una rilevazione svoltasi nel 2001 sul prezzo in
 Euro dei libri presenti nel catalogo di un importante editore
 italiano (n.420 osservazioni)
}
\examples{
head(libri, n = 10)
summary(libri)
}
\keyword{datasets}
