% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\docType{data}
\name{metabcg}
\alias{metabcg}
\title{METABCG data}
\format{A data.frame for meta-regression (Book, table
    20.1 pag 190)
\describe{
\item{study_name}{Study name}


}}
\source{
Borenstein et al. 2009 - Introduction to Meta-Analysis -
    Wiley, retrieved
    \href{here}{https://www.meta-analysis.com/pages/book_introduction_to_meta-analysis.php}
}
\usage{
metabcg
}
\description{
metabcg dataset.
}
\examples{
head(metabcg)
summary(metabcg)
}
\keyword{datasets}
