% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\docType{data}
\name{metacon}
\alias{metacon}
\title{METACON data}
\format{An object of class \code{data.frame} with 6 rows and 7 columns.}
\source{
Borenstein et al. 2009 - Introduction to Meta-Analysis -
    Wiley, retrieved
    \href{here}{https://www.meta-analysis.com/pages/book_introduction_to_meta-analysis.php}
}
\usage{
metacon
}
\description{
metacon dataset.
}
\examples{
head(metacon)
summary(metacon)
}
\keyword{datasets}
