% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\docType{data}
\name{metacor}
\alias{metacor}
\title{METACOR data}
\format{A data.frame for metaanalysis of correlation data (Book, table
    14.7 pag 98)
\describe{ \item{study}{Study name}
    \item{cor}{correlation} \item{n}{number of patients}

}}
\source{
Borenstein et al. 2009 - Introduction to Meta-Analysis -
    Wiley, retrieved
    \href{here}{https://www.meta-analysis.com/pages/book_introduction_to_meta-analysis.php}
}
\usage{
metacor
}
\description{
metacor dataset.
}
\examples{
head(metacor)
summary(metacor)
}
\keyword{datasets}
