% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mros.R
\docType{data}
\name{mros}
\alias{mros}
\title{Mros (sample) dataset}
\format{A data frame with 42 observations on the following variables.
 \describe{
   \item{\code{id}}{patient id}
   \item{\code{bmd}}{??}
   \item{\code{giage1}}{??}
   \item{\code{cutbmd}}{class of something}
   \item{\code{time}}{T}
   \item{\code{status}}{1 if ... 0 if ...}
}}
\usage{
mros
}
\description{
Mros (sample) dataset
}
\examples{
head(mros)
}
\keyword{datasets}
