% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msofa.R
\docType{data}
\name{msofa}
\alias{msofa}
\title{MSOFA data}
\format{A data.frame with XXX rows and YY variables: \describe{
    \item{c1}{Item C1}
    \item{c2}{Item C2}
    \item{c3}{Item C3}
    \item{c4}{Item C4}
    \item{c5}{Item C5}
    \item{c6}{Item C6}
    \item{c7}{Item C7}
    \item{c8}{Item C8}
    \item{c9}{Item C9}
    \item{c10}{Item C10}
    \item{c11}{Item C11}
    \item{c12}{Item C12}
    \item{c13}{Item C13}
    \item{c14}{Item C14}
    \item{c15}{Item C15}
    \item{c16}{Item C16}
    \item{c17}{Item c17}
    \item{c18}{Item C18}
    \item{c19}{Item C19}
    \item{c20}{Item C20}
    \item{age}{Age in years}
    \item{race}{Race}
    \item{marital}{Marital status}
    \item{educlevl}{Education level}
    \item{insured}{Insured}

}}
\source{
Pett, A.M, Lackey, N.R. and Sullivan J.J. (2003) Making
    sense of factor analysis, Sage
}
\usage{
msofa
}
\description{
msofa dataset.
}
\examples{
head(msofa, n = 10)
summary(msofa)
}
\keyword{datasets}
