% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo.R
\docType{data}
\name{produzione}
\alias{produzione}
\title{dataset PRODUZIONE}
\format{\describe{
\item{produzione}{indice mensile produzione}
}}
\source{
Piccolo, D., Statistica per le decisioni, Il Mulino 2010
}
\usage{
produzione
}
\description{
Indice mensile della Produzione industriale (corretto per i giorni
lavorativi) (Gennaio 1990 - Dicembre 2003)
}
\examples{
head(produzione, n = 10)
summary(produzione)
}
\keyword{datasets}
