% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piccolo_statistica_decisioni.R
\docType{data}
\name{studenti}
\alias{studenti}
\title{dataset STUDENTI}
\format{\describe{
\item{anno}{anno di rilevazione}
\item{genere}{genere dell'intervistato: studenti (M)=0; studentesse (F)=1}
\item{eta}{eta` in frazioni di anno (calcolato sulla base del mese/anno di nascita dichiarato)}
\item{peso}{peso (dichiarato) in kg}
\item{altezza}{altezza (dichiarata) in cm}
\item{dipsci}{possesso del diploma di maturita` scientifica: SI=1; No=0}
\item{dipcla}{possesso del diploma di maturita` classica: SI=1; No=0}
\item{diptec}{possesso del diploma di maturita` tecnica: SI=1; No=0}
\item{dipalt}{possesso di altri diploma di maturita`: SI=1; No=0}
\item{componenti}{numeri dei componenti della famiglia (compreso l'intervistato)}
\item{occhiali}{possesso di occhiali da vista: SI=1, NO=0}
\item{fumo}{abitudine al fumo: SI=1, NO=0}
\item{diploma}{diploma secondario dello studente}
}}
\source{
Piccolo, D., Statistica per le decisioni, Il Mulino 2010
}
\usage{
studenti
}
\description{
I dati riguardano n.2759 studenti presenti in aula durante la
 lezione di Statistica, svolta al I anno dei CdL in Scienze
 Politiche, Universita` di Napoli Federico II, e raccolti durante
 gli anni 1988-1995
}
\examples{
head(studenti, n = 10)
summary(studenti)
}
\keyword{datasets}
