% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaanalysis.R
\docType{data}
\name{metacon}
\alias{metacon}
\title{METACON data}
\format{A data.frame for metaanalysis of continuous data (Book, table 14.1
pag 93): \describe{
    \item{study}{Study name}
   \item{tm}{treated: mean}
   \item{tsd}{treated: standard deviation}
   \item{tn}{treated: total number of patients}
   \item{cm}{controls: mean}
   \item{csd}{controls: standard deviation}
   \item{cn}{controls: total number of patients}

}}
\source{
Borenstein et al. 2009 - Introduction to Meta-Analysis -
    Wiley, retrieved
    \href{here}{https://www.meta-analysis.com/pages/book_introduction_to_meta-analysis.php}
}
\usage{
metacon
}
\description{
metacon dataset.
}
\examples{
head(metacon)
summary(metacon)
}
\keyword{datasets}
